--- gov_starter_node-master/.gitignore
@@ -1,19 +1,0 @@
-.sass-cache
-.DS_Store
-.start.pid
-.port.tmp
-public
-assets/stylesheets/application.css
-lib/govuk_template.html
-govuk_modules
-node_modules/*
-.tmuxp.*
-.idea
-npm-debug.log
-build/*
-uploads/*
-test-results.xml
-build-info.json
-.env
-yarn-error.log
-.eslintcache
