--- gov_starter_node-master/assets/sass/style.sass
@@ -1,13 +1,0 @@
-// Path to assets for use with the file-url function
-// in the govuk frontend toolkit's url-helpers partial
-$path: "/public/images/"
-
-// Import GOV.UK elements from /govuk-modules/, this will import the frontend toolkit and some base styles.
-// Take a look in /govuk-modules/_govuk-elements.scss to see which files are imported.
-@import 'govuk-elements'
-
-#global-header #logo.hmppslogo
-  background: url('/public/images/logo.svg') no-repeat
-
-#global-header #logo .logo-img
-  margin-right: 10px
