--- gov_starter_node-master/log.js
@@ -1,8 +1,0 @@
-const bunyan = require('bunyan');
-const bunyanFormat = require('bunyan-format');
-
-const formatOut = bunyanFormat({ outputMode: 'short' });
-
-const log = bunyan.createLogger({ name: 'Starter app', stream: formatOut, level: 'debug' });
-
-module.exports = log;
