--- gov_starter_node-master/package.json
@@ -1,62 +1,0 @@
-{
-  "name": "gov-starter",
-  "version": "0.0.0",
-  "scripts": {
-    "start": "node server.js",
-    "start:dev": "yarn build && DEBUG=gov-starter-server* nodemon server.js | bunyan -o short",
-    "build": "yarn run css-build",
-    "css-build": "./bin/build-css",
-    "clean": "rm -rf public/* .port.tmp *.log build/* uploads/* test-results.xml",
-    "lint": "eslint . --cache --max-warnings 0",
-    "test": "NODE_ENV=test mocha test/*.spec.js test/**/*.spec.js"
-  },
-  "engines": {
-    "node": ">= 8.4.0",
-    "npm": ">=5.3.0"
-  },
-  "nodemonConfig": {
-    "ignore": [
-      ".circleci/*",
-      "migrations/*",
-      "node_modules/*",
-      "test/*"
-    ],
-    "delay": "2500"
-  },
-  "dependencies": {
-    "body-parser": "^1.18.2",
-    "bunyan": "^1.8.12",
-    "bunyan-format": "^0.2.1",
-    "bunyan-request-logger": "^1.1.0",
-    "compression": "^1.7.2",
-    "cookie-parser": "~1.4.3",
-    "cookie-session": "^2.0.0-beta.3",
-    "csurf": "^1.9.0",
-    "debug": "~3.1.0",
-    "dotenv": "^5.0.1",
-    "ejs": "~2.5.8",
-    "express": "~4.16.3",
-    "express-request-id": "^1.4.0",
-    "govuk-elements-sass": "^3.1.2",
-    "govuk_frontend_toolkit": "^7.4.2",
-    "govuk_template_jinja": "^0.23.1",
-    "helmet": "^3.12.0",
-    "http-errors": "~1.6.3",
-    "moment": "^2.22.1",
-    "morgan": "~1.9.0",
-    "node-sass-middleware": "0.11.0",
-    "passport": "^0.4.0",
-    "sass-middleware": "^0.0.3"
-  },
-  "devDependencies": {
-    "chai": "^4.1.2",
-    "chai-as-promised": "^7.1.1",
-    "chai-string": "^1.4.0",
-    "eslint": "^4.19.1",
-    "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-plugin-import": "^2.11.0",
-    "mocha": "^5.1.0",
-    "nodemon": "^1.17.3",
-    "sinon": "^4.5.0"
-  }
-}
