--- gov_starter_node-master/server/config.js
@@ -1,15 +1,0 @@
-const production = process.env.NODE_ENV === 'production';
-
-function get(name, fallback, options = {}) {
-  if (process.env[name]) {
-    return process.env[name];
-  }
-  if (fallback !== undefined && (!production || !options.requireInProduction)) {
-    return fallback;
-  }
-  throw new Error(`Missing env var ${name}`);
-}
-
-module.exports = {
-  sessionSecret: get('SESSION_SECRET', 'app-insecure-default-session', { requireInProduction: true }),
-};
