--- gov_starter_node-master/server/routes/index.js
@@ -1,15 +1,0 @@
-const express = require('express');
-
-module.exports = function Index({ logger, someService }) {
-  const router = express.Router();
-
-  router.get('/', (req, res) => {
-    logger.info('GET index');
-
-    const data = someService.getSomeData();
-
-    res.render('pages/index', { data });
-  });
-
-  return router;
-};
