--- gov_starter_node-master/server/routes/menu.js
@@ -1,13 +1,0 @@
-const express = require('express');
-
-module.exports = function ({logger, menuService}) {
-
-    const router = express.Router();
-
-    router.get('/menu', async (request, response) => {
-        logger.info('GET menu');
-        const data = await menuService.getMenuElement();
-        response.render('pages/index', {data})
-    });
-    return router;
-};
