--- gov_starter_node-master/server/services/menuService.js
@@ -1,22 +1,0 @@
-const contentClient = require('../data/contentClient');
-
-module.exports = function createMenuService() {
-
-    const getMenuElement = async () => {
-      const data = await contentClient.getMenu();
-      return  {
-        menu: formatMenu(data)
-      }
-    }
-
-    const formatMenu = result => {
-      const menuItems = result[0].below;
-      return Object.keys(menuItems).map(key => {
-          return menuItems[key].title;
-      });
-    }
-
-    return {
-        getMenuElement
-    }
-};
