--- gov_starter_node-master/server/services/someService.js
@@ -1,11 +1,0 @@
-module.exports = function createSomeService() {
-  function getSomeData() {
-    return {
-      text: 'Stuff from some service',
-    };
-  }
-
-  return {
-    getSomeData,
-  };
-};
