--- gov_starter_node-master/test/mocha.opts
@@ -1,5 +1,0 @@
---require ./test/.setup.js
---reporter spec
---colors
---bail
---timeout 25000
