+++ moj-node/.eslintignore
