+++ moj-node/assets/sass/style.sass
@@ -5,3 +5,9 @@ $path: "/public/images/"
 // Import GOV.UK elements from /govuk-modules/, this will import the frontend toolkit and some base styles.
 // Take a look in /govuk-modules/_govuk-elements.scss to see which files are imported.
 @import 'govuk-elements'
+
+#global-header #logo.hmppslogo
+  background: url('/public/images/logo.svg') no-repeat
+
+#global-header #logo .logo-img
+  margin-right: 10px
