+++ moj-node/log.js
@@ -1,7 +1,8 @@
 const bunyan = require('bunyan');
 const bunyanFormat = require('bunyan-format');
-const formatOut = bunyanFormat({outputMode: 'short'});
 
-const log = bunyan.createLogger({name: 'Starter app', stream: formatOut, level: 'debug'});
+const formatOut = bunyanFormat({ outputMode: 'short' });
+
+const log = bunyan.createLogger({ name: 'Starter app', stream: formatOut, level: 'debug' });
 
 module.exports = log;
