+++ moj-node/server.js
@@ -1,6 +1,8 @@
+require('dotenv').config();
+
 const app = require('./server/index');
 const log = require('./log');
 
-app.listen(app.get('port'), function() {
-    log.info('Licences server listening on port ' + app.get('port'));
+app.listen(app.get('port'), () => {
+  log.info(`Server listening on port ${app.get('port')}`);
 });
