+++ moj-node/server/routes/index.js
@@ -1,16 +1,15 @@
 const express = require('express');
 
-module.exports = function ({logger, someService, someOtherService}) {
-    const router = express.Router();
+module.exports = function Index({ logger, someService }) {
+  const router = express.Router();
 
-    router.get('/', (req, res) => {
-        // logger.info('GET index');
+  router.get('/', (req, res) => {
+    logger.info('GET index');
 
-        const data = someService.getSomeData();
-        const dataOther = someOtherService.getSomeData();
+    const data = someService.getSomeData();
 
-        res.render('pages/index', {data, dataOther})
-    });
+    res.render('pages/index', { data });
+  });
 
-    return router;
+  return router;
 };
