--- moj-node/server/services/someOtherService.js
@@ -1,12 +1,0 @@
-module.exports = function createSomeOtherService() {
-
-    function getSomeData() {
-        return {
-            other: 'All the things'
-        }
-    }
-
-    return {
-        getSomeData
-    }
-};
