+++ moj-node/server/services/someService.js
@@ -1,12 +1,11 @@
 module.exports = function createSomeService() {
-
-    function getSomeData() {
-        return {
-            text: 'Stuff from some service'
-        }
-    }
-    
+  function getSomeData() {
     return {
-        getSomeData
-    }
+      text: 'Stuff from some service',
+    };
+  }
+
+  return {
+    getSomeData,
+  };
 };
