--- moj-node/server/views/error.ejs
@@ -1,3 +1,0 @@
-<h1><%= message %></h1>
-<h2><%= error.status %></h2>
-<pre><%= error.stack %></pre>
