--- moj-node/server/views/index.ejs
@@ -1,75 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Starter App</title>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1"/>
-    <meta property="og:image" content="/public/images/opengraph-image.png?0.19.2"/>
-    <link rel='stylesheet' href='/public/stylesheets/application.css'/>
-    <link rel='shortcut icon' href='/public/images/favicon.ico?0.19.2' type="image/x-icon"/>
-    <link rel="mask-icon" href="/public/images/gov.uk_logotype_crown.svg?0.19.2" color="#0b0c0c"/>
-    <link rel="apple-touch-icon-precomposed" sizes="152x152" href="/public/images/apple-touch-icon-152x152.png?0.19.2"/>
-    <link rel="apple-touch-icon-precomposed" sizes="120x120" href="/public/images/apple-touch-icon-120x120.png?0.19.2"/>
-    <link rel="apple-touch-icon-precomposed" sizes="76x76" href="/public/images/apple-touch-icon-76x76.png?0.19.2"/>
-    <link rel="apple-touch-icon-precomposed" sizes="76x76" href="/public/images/apple-touch-icon-60x60.png?0.19.2"/>
-
-    <!--[if gt IE 8]><!-->
-    <link href="/public/stylesheets/govuk-template.css?" +version media="screen" rel="stylesheet" type="text/css"/>
-    <!--<![endif]-->
-    <!--[if gte IE 9]><!-->
-    <link href="/public/stylesheets/fonts.css?0.19.2" media="all" rel="stylesheet"/><!--<![endif]-->
-    <!--[if IE 8]>
-    <link href="/public/stylesheets/govuk-template-ie8.css?0.19.2" media="screen" rel="stylesheet"/><!--<![endif]-->
-
-    <!--[if lt IE 9]>
-    <link href="/public/stylesheets/grid-old-ie.css" media="screen" rel="stylesheet" type="text/css"/>
-    <link href="/public/stylesheets/application-oldie.css" media="screen" rel="stylesheet" type="text/css"/>
-    <script src="/public/javascripts/ie.js?0.19.2"></script>
-    <![endif]-->
-</head>
-<body>
-<header role="banner" id="global-header" class="with-proposition">
-    <div class="header-wrapper">
-        <div class="header-global">
-            <div class="header-logo">
-                <a href="/" title="Go to the homepage" id="logo" class="content hmppslogo">
-                    <img src="/public/images/hmpps_logo_small.png" width="54" height="48" alt=""/>
-                    <span>HMPPS</span>
-                </a>
-            </div>
-        </div>
-    </div>
-</header>
-
-<main class="app-container">
-    <ul>
-      <% for(var i=0; i<data.menu.length; i++) {%>
-        <li><%= data.menu[i] %></li>
-      <% } %>
-    </ul>
-</main>
-
-<footer class="group js-footer" id="footer" role="contentinfo">
-    <div class="footer-wrapper">
-        <div class="footer-meta">
-            <div class="footer-meta-inner">
-                <div class="open-government-licence">
-                    <p class="logo">
-                        <a href="https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/"
-                           rel="license">
-                            Open Government Licence
-                        </a>
-                    </p>
-                    <p>
-                        All content is available under the <a href="https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/"  rel="license">Open Government Licence v3.0</a>, except where otherwise stated
-                    </p>
-                </div>
-            </div>
-            <div class="copyright">
-                <a href="http://www.nationalarchives.gov.uk/information-management/re-using-public-sector-information/copyright-and-re-use/crown-copyright/">&copy; Crown copyright</a>
-            </div>
-        </div>
-    </div>
-</footer>
-</body>
-</html>
