+++ moj-node/server/views/pages/index.ejs
@@ -8,6 +8,11 @@
   <% include ../partials/header %>
 
   <main id="content" class="app-container" role="main">
+      <ul>
+        <% for(var i=0; i<data.menu.length; i++) {%>
+          <li><%= data.menu[i] %></li>
+        <% } %>
+      </ul>
     <div class="grid-row">
       <div class="column-two-thirds">
           <h1 class="heading-xlarge"><%= data.text %></h1>
