+++ moj-node/test/mocha.opts
@@ -1,3 +1,5 @@
---recursive
+--require ./test/.setup.js
 --reporter spec
---timeout 5000
+--colors
+--bail
+--timeout 25000
